requirements.txt
Flask==3.0.3
flask-cors==4.0.1
torch==2.3.0
transformers==4.41.1
accelerate==0.30.1
datasets[audio]==2.19.1
PyYAML==6.0.1
Flask-SQLAlchemy==3.1.1
sqlalchemy==2.0.30
pandas==2.2.2
huggingface_hub==0.23.2